Ensure necessary refs are fetched in GHA
