+++ src/semgrep_agent/utils.py
@@ -1,7 +1,10 @@
 import json
 import os
 import urllib.parse
+from contextlib import contextmanager
 from pathlib import Path
+from typing import cast
+from typing import Iterator
 from typing import List
 from typing import Optional
 from typing import TYPE_CHECKING
@@ -67,3 +70,74 @@ def zsplit(s: str) -> List[str]:
 
 def validate_publish_token(token: str) -> bool:
     return constants.PUBLISH_TOKEN_VALIDATOR.match(token) is not None
+
+
+@contextmanager
+def fix_head_for_github(
+    base_commit_ref: Optional[str] = None,
+    head_ref: Optional[str] = None,
+) -> Iterator[Optional[str]]:
+    """
+    GHA can checkout the incorrect commit for a PR (it will create a fake merge commit),
+    so we need to reset the head to the actual PR branch head before continuing.
+
+    Note that this code is written in a generic manner, so that it becomes a no-op when
+    the CI system has not artifically altered the HEAD ref.
+
+    :return: The baseline ref as a commit hash
+    """
+
+    stashed_rev: Optional[str] = None
+    base_ref: Optional[str] = base_commit_ref
+
+    if get_git_repo() is None:
+        yield base_ref
+        return
+
+    if base_ref:
+        # Preserve location of head^ after we possibly change location below
+        base_ref = git(["rev-parse", base_ref]).stdout.decode("utf-8").rstrip()
+
+    if head_ref:
+        stashed_rev = git(["branch", "--show-current"]).stdout.decode("utf-8").rstrip()
+        if not stashed_rev:
+            stashed_rev = git(["rev-parse", "HEAD"]).stdout.decode("utf-8").rstrip()
+        click.echo(f"| not on head ref {head_ref}; checking that out now...", err=True)
+        git.checkout([head_ref])
+
+    try:
+        if base_ref is not None:
+            click.echo("| scanning only the following commits:", err=True)
+            # fmt:off
+            log = git.log(["--oneline", "--graph", f"{base_ref}..HEAD"]).stdout  # type:ignore
+            # fmt: on
+            rr = cast(bytes, log).decode("utf-8").rstrip().split("\n")
+            r = "\n|   ".join(rr)
+            click.echo("|   " + r, err=True)
+
+        yield base_ref
+    finally:
+        if stashed_rev is not None:
+            click.echo(f"| returning to original head revision {stashed_rev}", err=True)
+            git.checkout([stashed_rev])
+
+
+def ensure_refs_fetched(meta: "GitMeta") -> None:
+    """Fetch context's head and base refs if they aren't already available.
+
+    This is needed e.g. when actions/checkout@v2 fetches only one commit.
+    """
+    if get_git_repo() is None:
+        return
+
+    refs = [meta.base_commit_ref, meta.head_ref]
+    for ref in refs:
+        if not ref:
+            continue
+        if git("cat-file", "commit", ref).exit_code == 0:
+            continue
+        click.echo(
+            f"| fetching locally unavialable ref '{ref}'",
+            err=True,
+        )
+        git.fetch("origin", ref)
